Associate user to post

+++ posts/views.py
@@ -13,6 +13,7 @@ def post_create(request):
 	form = PostForm(request.POST or None, request.FILES or None)
 	if form.is_valid():
 		instance = form.save(commit=False)
+		instance.user = request.user
 		instance.save()
 		messages.success(request, 'Successfully Created.')
 		return HttpResponseRedirect(instance.get_absolute_url())
@@ -54,9 +55,12 @@ def post_update(request, slug=None):
 	if not (request.user.is_staff or request.user.is_superuser):
 		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
+	if instance.user != request.user:
+		raise Http404
 	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 	if form.is_valid():
 		instance = form.save(commit=False)
+		instance.user = request.user
 		instance.save()
 		messages.success(request, 'Item Saved.')
 		return HttpResponseRedirect(instance.get_absolute_url())
@@ -72,6 +76,8 @@ def post_delete(request, slug=None):
 	if not (request.user.is_staff or request.user.is_superuser):
 		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
+	if instance.user != request.user:
+		raise Http404
 	instance.delete()
 	messages.success(request, 'Successfully Deleted.')
 	return redirect('posts:list')
